remove controversial note
I decided to purely remove this note.
In a discussion with @doerthe, we agreed that D-derivation could be handled
- in the concrete-to-abstract syntax mapping (by normalizing literals)
  but I don't think that should be a requirement in general
- in some other semantic extensions (in particular math-entailment)
  but those are not presented currently in the document.

RDF simple entailment does not deal with datatypes,
neither does N3 base entailment.
I don't think we need to apologize for that.

We can add a note about the relationship with RDF's D-entailment
when we introduce datatype support in Math Semantics.